conventional folderstructure, cleanup, eventemitter added
