+++ ProjectDaden/src/app/app.component.ts
@@ -1,8 +1,8 @@
 import { Component } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
-import { NavSidebarComponent } from './nav-sidebar/nav-sidebar.component';
-import { AsidePreviewsComponent } from './aside-previews/aside-previews.component';
-import { ToolColorPickerComponent } from './tool-color-picker/tool-color-picker.component';
+import { NavSidebarComponent } from './features/nav-sidebar/nav-sidebar.component';
+import { AsidePreviewsComponent } from './features/aside-previews/aside-previews.component';
+import { ToolColorPickerComponent } from './features/tool-color-picker/tool-color-picker.component';
 
 @Component({
   selector: 'app-root',
