+++ ProjectDaden/src/app/shared/components/daden-dropdown/daden-dropdown.component.spec.ts
@@ -1,18 +1,18 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { DropdownComponent } from './dropdown.component';
+import { DadenDropdownComponent } from './daden-dropdown.component';
 
-describe('DropdownComponent', () => {
-  let component: DropdownComponent;
-  let fixture: ComponentFixture<DropdownComponent>;
+describe('DadenDropdownComponent', () => {
+  let component: DadenDropdownComponent;
+  let fixture: ComponentFixture<DadenDropdownComponent>;
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [DropdownComponent]
+      imports: [DadenDropdownComponent]
     })
     .compileComponents();
 
-    fixture = TestBed.createComponent(DropdownComponent);
+    fixture = TestBed.createComponent(DadenDropdownComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
